08_a2a_dual_agents/langgraph_agent_server/pyproject.toml
[project]
name = "a2a-langgraph-agent"
version = "0.1.0"
description = "LangGraph-based agent server for A2A orchestration demo"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "langgraph>=0.2.30",
  "langchain-openai>=0.1.17",
  "python-dotenv>=1.0",
]

[tool.uv]
package = false
